FEAT: Use less intrusive colours in footer
Changed the background colour of the links to the same colour as
 ordinary text.
Changed the text colour and colour on hover to match.
Added padding around the link text.
Moved the copyright line away from the text above it.
Added a border between the two link blocks on the right hand side to
 give them some separation.