BUG: "Therapies" menu not aligned with other menus
In the desktop version, the "Therapies" menu is not aligned with the
 other menus.
Changed it from a <span> element to an <a> element.
This caused an issue with the mobile version of the menu.
Now clicking on the Therapies menu item in the mobile menu goes
 straight to the linked page, rather than displaying a dropdown menu
 showing Therapy options.
Fixed this by setting the pointer-events to 'none' for this anchor on
 mobile devices, and reset the pointer-events to work for desktop.