FIX: Submenu disappearing when moving cursor on it
Bug 1: When moving from the menu heading to the submenu, the submenu
 would sometimes disappear.
This was caused by a gap between the menu heading and the submenu.
When the mouse cursor hit the gap, the submenu would disappear.
This gap has been removed.
Also removed the triangular "arrow" from the mobile version.
Bug 2: On the destop version, hovering over a submenu item would
 highlight the item, but the highlight would only highlight the content
 not the entire width of the submenu.
Now when hovering over a submenu item the entire line is highlighted.
Also added borders to delineate between the submenu items.